pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "inli-rental-scraper"
version = "0.1.0"
description = "AI-powered web scraper for Inli rental listings"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "KevinT", email = "kevin.koffi.tanoh@gmail.com"},
]
dependencies = [
    "browser-use>=0.9.0",
    "fastapi[standard]>=0.121.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
inli-scraper = "agent:main"

